services/rare-identity-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rare-identity-core"
version = "0.2.0"
description = "Rare Identity Core API service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "pydantic>=2.8.0",
  "httpx>=0.27.0",
  "rare-identity-protocol>=0.1.0",
  "rare-identity-verifier>=0.1.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2.0",
]
storage = [
  "psycopg[binary]>=3.2.0",
  "redis>=5.0.0",
]
gcp = [
  "google-cloud-kms>=3.0.0",
  "google-cloud-secret-manager>=2.21.0",
]
serve = [
  "uvicorn[standard]>=0.34.0",
  "dnspython>=2.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["services"]
include = ["rare_api*"]

[tool.setuptools.package-data]
rare_api = ["static/*"]
